feat: add flag for distributed queries with overlapping intervals
