+++ test/e2e/query_test.go
@@ -2376,93 +2376,3 @@ func TestDistributedEngineWithExtendedFunctions(t *testing.T) {
 	}, time.Now, promclient.QueryOptions{}, 1)
 	testutil.Equals(t, model.SampleValue(0), result[0].Value)
 }
-
-func TestDistributedEngineWithDisjointTSDBs(t *testing.T) {
-	t.Skip("skipping test as this replicates a bug")
-	e, err := e2e.New(e2e.WithName("dist-disj-tsdbs"))
-	testutil.Ok(t, err)
-	t.Cleanup(e2ethanos.CleanScenario(t, e))
-
-	ctx := context.Background()
-	l := log.NewLogfmtLogger(os.Stdout)
-	now := time.Now()
-
-	bucket1 := "dist-disj-tsdbs-test1"
-	minio1 := e2edb.NewMinio(e, "1", bucket1, e2edb.WithMinioTLS())
-	testutil.Ok(t, e2e.StartAndWaitReady(minio1))
-
-	bkt1, err := s3.NewBucketWithConfig(l, e2ethanos.NewS3Config(bucket1, minio1.Endpoint("http"), minio1.Dir()), "test", nil)
-	testutil.Ok(t, err)
-
-	// Setup a storage GW with 2 blocks that have a gap to trigger distributed query MinT bug
-	dir1 := filepath.Join(e.SharedDir(), "tmp1")
-	testutil.Ok(t, os.MkdirAll(filepath.Join(e.SharedDir(), dir1), os.ModePerm))
-	blockID1, err := e2eutil.CreateBlockWithBlockDelay(ctx,
-		dir1,
-		[]labels.Labels{labels.FromStrings("__name__", "foo", "instance", "foo_1")},
-		1000,
-		timestamp.FromTime(now.Add(-10*time.Hour)),
-		timestamp.FromTime(now.Add(-8*time.Hour)),
-		30*time.Minute,
-		labels.FromStrings("prometheus", "p1", "replica", "0"),
-		0,
-		metadata.NoneFunc,
-	)
-	testutil.Ok(t, err)
-	testutil.Ok(t, objstore.UploadDir(ctx, l, bkt1, path.Join(dir1, blockID1.String()), blockID1.String()))
-
-	blockID2, err := e2eutil.CreateBlockWithBlockDelay(ctx,
-		dir1,
-		[]labels.Labels{labels.FromStrings("__name__", "foo", "instance", "foo_1")},
-		1000,
-		timestamp.FromTime(now.Add(-4*time.Hour)),
-		timestamp.FromTime(now.Add(-2*time.Hour)),
-		30*time.Minute,
-		labels.FromStrings("prometheus", "p1", "replica", "0"),
-		0,
-		metadata.NoneFunc,
-	)
-	testutil.Ok(t, err)
-	testutil.Ok(t, objstore.UploadDir(ctx, l, bkt1, path.Join(dir1, blockID2.String()), blockID2.String()))
-	store1 := e2ethanos.NewStoreGW(
-		e,
-		"s1",
-		client.BucketConfig{
-			Type:   client.S3,
-			Config: e2ethanos.NewS3Config(bucket1, minio1.InternalEndpoint("http"), minio1.InternalDir()),
-		},
-		"",
-		"",
-		nil,
-	)
-	testutil.Ok(t, e2e.StartAndWaitReady(store1))
-
-	querierLeaf1 := e2ethanos.NewQuerierBuilder(e, "1", store1.InternalEndpoint("grpc")).Init()
-
-	// We need another querier to circumvent the passthrough optimizer
-	promConfig2 := e2ethanos.DefaultPromConfig("p2", 0, "", "", e2ethanos.LocalPrometheusTarget)
-	prom2, sidecar2 := e2ethanos.NewPrometheusWithSidecar(e, "p2", promConfig2, "", e2ethanos.DefaultPrometheusImage(), "")
-	querierLeaf2 := e2ethanos.NewQuerierBuilder(e, "2", sidecar2.InternalEndpoint("grpc")).Init()
-
-	querierDistributed := e2ethanos.NewQuerierBuilder(e, "3",
-		querierLeaf1.InternalEndpoint("grpc"),
-		querierLeaf2.InternalEndpoint("grpc"),
-	).
-		WithEngine(v1.PromqlEngineThanos).
-		WithQueryMode("distributed").
-		Init()
-
-	testutil.Ok(t, e2e.StartAndWaitReady(querierLeaf1, prom2, sidecar2, querierLeaf2, querierDistributed))
-
-	// We would expect 2x2h ranges for the 2 blocks containing foo samples. That would be around 240 expected sample pairs in the result matrix.
-	// We assert on more then 200 to reduce flakiness
-	rangeQuery(t, ctx, querierDistributed.Endpoint("http"), func() string { return "foo" }, timestamp.FromTime(now.Add(-24*time.Hour)), timestamp.FromTime(now), 60, promclient.QueryOptions{}, func(res model.Matrix) error {
-		if res.Len() < 1 {
-			return errors.New("No result series returned")
-		}
-		if nvals := len(res[0].Values); nvals < 200 {
-			return errors.Errorf("Too few values in result matrix, got %d, expected > 200", nvals)
-		}
-		return nil
-	})
-}
